one cmd excutable all good, still need to close all opened fds...
