+++ Mandatory/inc/minishell.h
@@ -6,7 +6,7 @@
 /*   By: otaraki <otaraki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/22 01:12:28 by ebennix           #+#    #+#             */
-/*   Updated: 2023/08/31 23:59:09 by otaraki          ###   ########.fr       */
+/*   Updated: 2023/09/01 01:36:39 by otaraki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -82,6 +82,7 @@ void	my_echo(char **av, t_env **env);
 void	me_pwd(char **av, t_env **env);
 void	export_it(char **av, t_env **env);
 void	exceute_it(t_token **data, t_env **env);
+int 	red_open(t_type red, char *f_name);
 char	*get_next_line(int fd);
 
 // void tokeni(char *f_line);
